backend/requirements.txt
pyramid
waitress
pyramid_tm
SQLAlchemy
pyramid_debugtoolbar
pyramid_retry
webtest
PyJWT